csrc: post-C4 residual cleanup pass
Three independent cleanups landing together (shared files, distinct
hunks). Build clean; smoke + profile green at production shape
(world=16, num_sms=64, T=8192, K=4): 6.57 ms/iter, combine_main_kernel
600 µs, dispatch_main_kernel 480 µs. Per-kernel medians within
per-allocation Slurm variance of recent baselines.

1. Drop dead `get_*_clean_meta` helpers
─────────────────────────────────────────
After C4 the two helpers in `api.cuh` (`get_rdma_clean_meta` /
`get_nvl_clean_meta`) returned `count=0` for all production configs —
the dispatch/combine cleanup kernels they sized were deleted in C4 and
no call site remained. ~40 LOC removed plus one stale reference in
`internode.cu`'s top-of-file comment. Pure deletion.

2. Combine NVL `nvl_channel_head/tail` on the C4 gen-stamp protocol
───────────────────────────────────────────────────────────────────
Closes the last cleanup-shaped protection gap left after a27ee02. The
combine sender's `cached_channel_tail_idx` and the coordinator's
`last_nvl_head[i]` were iter-local (start at 0 each iter), relying on
`cached_notify_combine_kernel` block 0's bulk memset to wipe inter-iter
residue. C4 retired block 0; the receiver-side spin
`cached_nvl_channel_tail_idx <= expected_head` could then exit early on
prior-iter residue (e.g. iter-N's tail = 1000, iter-N+1's expected_head
= 5 → exit immediately, read uninitialized ring slot). Equivalent
hazard on `nvl_channel_head` for the sender's backpressure check.

Fix mirrors dispatch's C4 layout:
- Sender (kNVLSender) M1 force-writes `nvl_channel_tail` with
  `nvl_pack(combine_seq, 0)` at warp entry, one IPC store per
  (channel, dst_nvl, rdma_src) lane via `lane_id < kNumRDMARanks`.
- Sender's per-chunk tail writes pack `cached_channel_tail_idx` with
  the same seq.
- Sender's head reads now filter via `nvl_seq_match(raw, combine_seq)`;
  pre-tag residue leaves `cached_channel_head_idx` at its last-valid
  value (initial 0) until a matching read arrives.
- Forwarder's `nvl_channel_tail` spin reads similarly seq-match.
- Coordinator (is_forwarder_sm branch) M1 force-writes
  `nvl_channel_head` at entry, one IPC store per (dst_nvl, rdma_src)
  slot in the channel; its later progress writes pack `min_head` with
  the same seq.

`combine_seq` is in scope (already a kernel param) and identical to
`dispatch_seq` per `stream_moe.py:262, 297`. Combine NVL slots are
written only by `combine_main_kernel`, so there's no cross-kernel
seq-space aliasing.

3. Revert M2's `reader_prev_*` int64 → uint32 (wrap-safe CAS)
─────────────────────────────────────────────────────────────
M2 widened `dispatch_reader_prev_*` / `combine_reader_prev_*` to int64
to track cumulative-since-Buffer-creation across the 4-byte NIC AMO's
2^32 wrap (M1+M2 commit 60ec765). Opt1 (6403b42) kept the int64 but
already did modular uint32 subtraction at the read side — the int64
high bits were only carried for the writeback's `atomicMax<unsigned
long long>` cross-warp ordering, never read back as wider than uint32.

The simpler design: store only `uint32_t` (the slot's low 32 bits) in
`reader_prev_*`. Read-side math stays modular: `(slot_low - prev_low)
mod 2^32` — works through arbitrary wraps because adjacent-iter advance
≪ 2^32 (production: ~128 tokens/iter/slot). Write-side uses a
signed-difference CAS helper: pick the candidate that is "ahead" by
signed int32 arithmetic, which handles the cross-iter wrap boundary
correctly (plain `atomicMax<uint32>` would prefer pre-wrap 2^32-1 over
post-wrap 0, breaking iter-to-iter ordering — `(int32_t)(cur - old) > 0`
picks the temporally-latest regardless of wrap edge).

Within an iter, the slot is monotone (per-iter advance ≪ 2^32 so no
intra-iter wrap), so the CAS loop converges in at most a few retries
across the ~7 contending warps per (channel, peer). Across iters, the
writeback is stream-ordered against the next kernel's read — no race.

Wrap horizon: effectively infinite. The protocol no longer compares
across-iter slot values directly; the array storing those values is
allowed to cycle through 2^32 every ~500k training steps without
correctness impact.

Diff vs Opt1: helper simplifies from `int64 + uint32` reconstruction
to a pure CAS loop on uint32. 6 dispatch + 3 combine call sites drop
the `prev_*_at_entry` second argument. Memory: 4 arrays × 4× smaller
(~1 KB total vs ~2 KB; both tiny). LOC: net ~30 lines simpler.

Smoke + profile evidence the uint32 revert is perf-neutral
(kernel_y_bwd unchanged at 1022 µs, vs Opt1's 1023 µs at the same
profile shape). The +200 µs gap from C4 → M1+M2 → Opt1 lives elsewhere
and disproves the int64-widening hypothesis. Next step: investigate
M1's iter-entry force-write of `nvl_channel_head/tail` in
`dispatch_main_kernel` / `dispatch_grads_main_kernel` as the candidate
source of the gap.

+++ csrc/deep_ep.hpp
@@ -192,16 +192,19 @@ private:
     // amos), unlike an in-kernel `ld(slot)` seed which is racy because
     // peer amos land asynchronously through the NIC.
     //
-    // Sized [max_num_channels × num_rdma_ranks] int32, ~512 B each at
-    // production; ~2 KB total. Allocated zero-init in `Buffer::Buffer`,
-    // freed in destructor.
-    // int64_t (M2): the NIC AMOs that drive these counters were widened from
-    // 4-byte to 8-byte (mlx5 MLX5_OPCODE_ATOMIC_FA); int32 wraps at ~250k
-    // training steps × 32 layers, int64 horizon is effectively infinite.
-    int64_t* dispatch_reader_prev_head = nullptr;
-    int64_t* dispatch_reader_prev_tail = nullptr;
-    int64_t* combine_reader_prev_head  = nullptr;
-    int64_t* combine_reader_prev_tail  = nullptr;
+    // Sized [max_num_channels × num_rdma_ranks] uint32, ~256 B each at
+    // production; ~1 KB total. Allocated zero-init in `Buffer::Buffer`,
+    // freed in destructor. Storage is uint32 (matching the on-the-wire
+    // 4-byte AMO width); cross-iter wrap is absorbed by modular uint32
+    // subtraction in the read-side `cur - prev` math and a signed-difference
+    // CAS in the write-side helper (`atomicmax_reader_prev_cumulative`).
+    // Wrap horizon is effectively infinite — the protocol never compares
+    // across-iter slot values directly, so it doesn't care that the slot
+    // (and the array) cycle through 2^32 every ~500k training steps.
+    uint32_t* dispatch_reader_prev_head = nullptr;
+    uint32_t* dispatch_reader_prev_tail = nullptr;
+    uint32_t* combine_reader_prev_head  = nullptr;
+    uint32_t* combine_reader_prev_tail  = nullptr;
     // Persistent prev-sentinel array for the RDMA dispatch meta region (C3).
     // The meta SymBuffer's slot 30 ("kRdmaMetaSentinelSlot") accumulates
     // across iters via amo_nonfetch_add — sender bulk_puts the 18 data ints
